Revert "Restrict checking the number of arguments."
This reverts commit 122a7bccaff11def2c12cfccdd00244394ed3478.
